image: the dumper takes the sorted containers, entries and all
A set and a struct are sorted arrays inline in region pages, so the copy
takes them in order and never re-sorts: every key and element an image can
carry ranks by its own content, so the hydrating instance's comparator agrees
with the order the dump wrote. A struct's keys go through the value walk like
any other value, so a string or array key points inside the image, and a
symbol or keyword key leaves its spelling in the name table — keys reach the
walk nowhere else.

An entry cannot be copied. Its key is a `repr(Rust)` enum with padding of its
own, so the dumper assembles an entry from probed extents exactly as it
assembles an object slot. The probe is now written once against a trait and
answers for both records, and the fingerprint carries the key variants and
the entry's two offsets beside the heap ones.

The probe found the layout fact this rests on: a `Bool` key's payload sits in
the byte after the discriminant, not at the next word. So the discriminant
span is a property of the record — eight bytes for a `HeapObject`, one for a
`TableKey` — rather than the constant it was.

The verifier measures a struct's extent in entries. `TableKey` grows the
`heap_value`/`with_heap_value` pair, so which arms hold a `Value` is said in
one place.

Three files past the reading budget are split, each along a seam it already
had: the copying walk leaves the dumper, the layout probe splits by the
record it measures, and `value/types.rs` gives the key model and the sorted
lookup files of their own, re-exporting both so no import moves.
