--- src/image/dump.rs
@@ -1,379 +1,0 @@
-// audited: 2026-09-08
-//! The dumper: a compacting copy of a sealed data graph into a scratch
-//! region, written out as an image file.
-//!
-//! docs/impl/image.md
-//! docs/impl/image/format.md
-//!
-//! Beside the pages the file carries relocations, a page table, an object
-//! index, and the spellings of the symbols and keywords in the body. The
-//! store milestone's scope is data-only: pairs, strings, bytes, arrays,
-//! floats, and the portable immediates (ints, inline floats, bools, nil, the
-//! empty list, symbols and keywords — both payloads are stable name hashes).
-//! Anything else fails the dump with an error naming the variant.
-//!
-//! Determinism is engineered: the copy visits children in order, the visited
-//! map is only ever probed (never iterated), and the file's page bytes are
-//! assembled from a zeroed buffer. Object slots are never copied wholesale —
-//! each receives only its discriminant byte and probed leaf-field extents
-//! (`layout::write_canonical`), so neither recycled-page residue nor a
-//! construction temporary's uninitialized padding can reach the artifact,
-//! and two dumps of the same graph are byte-identical whole files.
-
-use std::collections::{BTreeSet, HashMap};
-use std::io::Write;
-use std::mem::size_of;
-use std::path::Path;
-
-use crate::hir::region::RuntimeRegion;
-use crate::symbol::SymbolTable;
-use crate::value::fiberheap::FiberHeap;
-use crate::value::heap::{deref, HeapObject, Pair};
-use crate::value::region_slice::RegionSlice;
-use crate::value::repr::{
-    TAG_EMPTY_LIST, TAG_FALSE, TAG_FLOAT, TAG_INT, TAG_KEYWORD, TAG_NIL, TAG_SYMBOL, TAG_TRUE,
-};
-use crate::value::Value;
-
-use super::format::{self, Header, PageEntry};
-use super::layout;
-use super::ImageError;
-
-/// Dump `root`'s value graph to `path`, atomically (temp file + rename).
-/// The graph must be sealed data; a refused value fails the dump before any
-/// byte is written, and the scratch region is dropped either way.
-///
-/// `symbols` is the dumping instance's display memo. Every symbol and keyword
-/// the walk meets takes its spelling from it into the image's name table; a
-/// spelling this instance never learned is simply absent.
-pub fn dump(
-    heap: &mut FiberHeap,
-    symbols: &SymbolTable,
-    root: Value,
-    path: &Path,
-) -> Result<(), ImageError> {
-    let scratch = heap.new_runtime_region();
-    let result = dump_into(heap, scratch, symbols, root, path);
-    heap.decref_region_if_present(scratch);
-    result
-}
-
-fn dump_into(
-    heap: &mut FiberHeap,
-    scratch: RuntimeRegion,
-    symbols: &SymbolTable,
-    root: Value,
-    path: &Path,
-) -> Result<(), ImageError> {
-    let mut walk = Walk::new(symbols);
-    let copied = copy_value(heap, scratch, root, &mut walk)?;
-    let names = walk.into_name_table();
-
-    // Page layout, largest page first: packing in descending size order keeps
-    // every page's offset a multiple of its own size (§ File format).
-    let mut layouts = heap
-        .region_pool(scratch)
-        .map(|p| p.page_layouts())
-        .unwrap_or_default();
-    layouts.sort_by_key(|l| std::cmp::Reverse(l.len));
-    let mut intervals: Vec<(usize, usize, u64)> = Vec::new(); // (base, len, rel)
-    let mut rel = 0u64;
-    for l in &layouts {
-        intervals.push((l.base, l.len, rel));
-        rel += l.len as u64;
-    }
-    let pages_len = rel;
-    let off_of = |addr: usize| -> Option<u64> {
-        intervals
-            .iter()
-            .find(|&&(base, len, _)| addr >= base && addr < base + len)
-            .map(|&(base, _, r)| r + (addr - base) as u64)
-    };
-    let in_image = |off: Option<u64>| -> Result<u64, ImageError> {
-        off.ok_or_else(|| {
-            ImageError::Corrupt("dump: a copied value points outside the scratch region".into())
-        })
-    };
-
-    // Walk the copied objects: canonical slot bytes into the zeroed pages
-    // buffer, index entries, relocation slots, and the meaningful data
-    // spans (slice backings) to copy into the file.
-    let mut pages = vec![0u8; pages_len as usize];
-    let mut relocs: Vec<(u64, u64)> = Vec::new();
-    let mut index: Vec<(u64, u64)> = Vec::new();
-    let mut data_spans: Vec<(u64, usize, usize)> = Vec::new(); // (rel, src, len)
-    if let Some(pool) = heap.region_pool(scratch) {
-        for obj in pool.live_objects() {
-            let addr = obj as *const HeapObject as usize;
-            let obj_off = in_image(off_of(addr))?;
-            index.push((obj_off, obj.tag() as u64));
-            let dst = obj_off as usize;
-            layout::write_canonical(obj, &mut pages[dst..dst + size_of::<HeapObject>()]);
-            let mut slot = |slot_addr: usize, target: usize| -> Result<(), ImageError> {
-                let s = in_image(off_of(slot_addr))?;
-                let t = in_image(off_of(target))?;
-                relocs.push((s, t));
-                Ok(())
-            };
-            match obj {
-                HeapObject::Pair(pair) => {
-                    for v in [&pair.first, &pair.rest] {
-                        if let Some(p) = v.as_heap_ptr() {
-                            slot(&v.payload as *const u64 as usize, p as usize)?;
-                        }
-                    }
-                }
-                HeapObject::LString { s, .. } => {
-                    slice_slots(s, &mut slot, &mut data_spans, &off_of)?;
-                }
-                HeapObject::LBytes { data, .. } => {
-                    slice_slots(data, &mut slot, &mut data_spans, &off_of)?;
-                }
-                HeapObject::LArray { elements, .. } => {
-                    slice_slots(elements, &mut slot, &mut data_spans, &off_of)?;
-                    for v in elements.iter() {
-                        if let Some(p) = v.as_heap_ptr() {
-                            slot(&v.payload as *const u64 as usize, p as usize)?;
-                        }
-                    }
-                }
-                HeapObject::Float(_) => {}
-                other => {
-                    // copy_value only allocates the variants above.
-                    unreachable!("unexpected {:?} in dump scratch", other.tag());
-                }
-            }
-        }
-    }
-    relocs.sort_unstable();
-    index.sort_unstable();
-
-    // Object slots are already canonical in `pages`; add the slice
-    // backings. Headers, gaps, alignment slack, and slot padding stay zero
-    // (see the module docs on determinism).
-    let entries: Vec<PageEntry> = layouts
-        .iter()
-        .map(|l| PageEntry {
-            size: l.len as u64,
-            obj_cursor: l.obj_cursor as u64,
-            data_cursor: l.data_cursor as u64,
-        })
-        .collect();
-    for &(rel, src, len) in &data_spans {
-        let bytes = unsafe { std::slice::from_raw_parts(src as *const u8, len) };
-        pages[rel as usize..rel as usize + len].copy_from_slice(bytes);
-    }
-    // Canonicalize every relocation slot to zero: its dump-time content is a
-    // scratch-region absolute address — meaningless to the file and rewritten
-    // wholesale by hydration's relocation pass.
-    for &(slot, _) in &relocs {
-        pages[slot as usize..slot as usize + 8].fill(0);
-    }
-
-    let (root_is_heap, root_payload) = match copied.as_heap_ptr() {
-        Some(p) => (true, in_image(off_of(p as usize))?),
-        None => (false, copied.payload),
-    };
-    let mut name_table = Vec::new();
-    for name in &names {
-        format::write_name(&mut name_table, name);
-    }
-
-    let header = Header {
-        pages_len,
-        n_pages: entries.len() as u64,
-        n_relocs: relocs.len() as u64,
-        n_objects: index.len() as u64,
-        root_tag: copied.tag,
-        root_payload,
-        root_is_heap,
-        names_len: name_table.len() as u64,
-        fingerprint: format::fingerprint(),
-    };
-
-    let mut file_bytes = header.to_block()?;
-    file_bytes.extend_from_slice(&pages);
-    for e in &entries {
-        format::write_page_entry(&mut file_bytes, *e);
-    }
-    for &(s, t) in &relocs {
-        format::write_u64_pair(&mut file_bytes, s, t);
-    }
-    for &(o, t) in &index {
-        format::write_u64_pair(&mut file_bytes, o, t);
-    }
-    file_bytes.extend_from_slice(&name_table);
-    debug_assert_eq!(file_bytes.len() % 8, 0);
-    debug_assert!(format::pages_offset() as u64 + pages_len <= file_bytes.len() as u64);
-
-    // Never rewrite an image file in place (§ Hydration): a mapped old inode
-    // must stay stable while a new image replaces the path.
-    let tmp = path.with_file_name(format!(
-        ".{}.tmp{}",
-        path.file_name().and_then(|n| n.to_str()).unwrap_or("image"),
-        std::process::id()
-    ));
-    let mut f = std::fs::File::create(&tmp)?;
-    f.write_all(&file_bytes)?;
-    f.sync_all()?;
-    drop(f);
-    if let Err(e) = std::fs::rename(&tmp, path) {
-        let _ = std::fs::remove_file(&tmp);
-        return Err(e.into());
-    }
-    Ok(())
-}
-
-/// Record a slice field's relocation slot (its `ptr`, at offset 0 of the
-/// `repr(C)` `RegionSlice`) and its backing bytes as a data span. An empty
-/// slice has a dangling constant pointer — no slot, no span.
-fn slice_slots<T: 'static>(
-    s: &RegionSlice<T>,
-    slot: &mut impl FnMut(usize, usize) -> Result<(), ImageError>,
-    data_spans: &mut Vec<(u64, usize, usize)>,
-    off_of: &impl Fn(usize) -> Option<u64>,
-) -> Result<(), ImageError> {
-    if s.is_empty() {
-        return Ok(());
-    }
-    let backing = s.as_ptr() as usize;
-    slot(s as *const RegionSlice<T> as usize, backing)?;
-    let len = s.len() * std::mem::size_of::<T>();
-    let rel = off_of(backing).ok_or_else(|| {
-        ImageError::Corrupt("dump: a slice backing lies outside the scratch region".into())
-    })?;
-    data_spans.push((rel, backing, len));
-    Ok(())
-}
-
-/// What the copying walk carries: the sharing map, and the spellings met so
-/// far. Both are per-dump state the recursion threads through every value.
-struct Walk<'a> {
-    /// Source payload address → its copy in the scratch region.
-    visited: HashMap<usize, Value>,
-    /// The dumping instance's display memo, read for spellings.
-    memo: &'a SymbolTable,
-    /// The spellings met, deduplicated and ordered by name — the order the
-    /// name table is written in, so one graph writes one table whatever order
-    /// the memo learned them in.
-    names: BTreeSet<Box<str>>,
-}
-
-impl<'a> Walk<'a> {
-    fn new(memo: &'a SymbolTable) -> Self {
-        Walk {
-            visited: HashMap::new(),
-            memo,
-            names: BTreeSet::new(),
-        }
-    }
-
-    /// Record the spelling of a symbol or keyword the walk just met. A value
-    /// whose spelling this instance never learned contributes none: it still
-    /// dumps, and it still prints as `#<symbol:hash>` on the other side
-    /// (docs/impl/symbol.md).
-    fn note_name(&mut self, v: Value) {
-        let memo = self.memo;
-        let name = match v.as_symbol() {
-            Some(id) => memo.name(id),
-            None => crate::value::keyword::resolve_keyword_name(Some(memo), v.payload),
-        };
-        if let Some(name) = name {
-            if !self.names.contains(name) {
-                self.names.insert(name.into());
-            }
-        }
-    }
-
-    fn into_name_table(self) -> Vec<Box<str>> {
-        self.names.into_iter().collect()
-    }
-}
-
-/// Deep-copy one sealed data value into the scratch region, preserving
-/// sharing through the walk's visited map (keyed on source payload address).
-/// A value outside the spike's sealed set fails the copy, naming the variant.
-fn copy_value(
-    heap: &mut FiberHeap,
-    region: RuntimeRegion,
-    v: Value,
-    walk: &mut Walk,
-) -> Result<Value, ImageError> {
-    if !v.is_heap() {
-        return match v.tag {
-            TAG_INT | TAG_FLOAT | TAG_NIL | TAG_TRUE | TAG_FALSE | TAG_EMPTY_LIST => Ok(v),
-            // A symbol and a keyword payload are name hashes, so the value
-            // means the same thing in every process; only the spelling has to
-            // travel beside it.
-            TAG_SYMBOL | TAG_KEYWORD => {
-                walk.note_name(v);
-                Ok(v)
-            }
-            _ => Err(ImageError::Unsupported(format!(
-                "immediate {} is not portable data",
-                v.type_name()
-            ))),
-        };
-    }
-    let key = v.payload as usize;
-    if let Some(&copy) = walk.visited.get(&key) {
-        return Ok(copy);
-    }
-    let obj = unsafe { deref(v) };
-    if obj.traits() != Value::NIL {
-        return Err(ImageError::Unsupported(format!(
-            "{:?} carries traits — instance state the image cannot own",
-            obj.tag()
-        )));
-    }
-    let copy = match obj {
-        HeapObject::Pair(pair) => {
-            let first = copy_value(heap, region, pair.first, walk)?;
-            let rest = copy_value(heap, region, pair.rest, walk)?;
-            heap.alloc_in_region(HeapObject::Pair(Pair::new(first, rest)), region)
-        }
-        HeapObject::LString { s, .. } => {
-            let slice = heap.alloc_region_slice_in_region(s.as_slice(), region);
-            heap.alloc_in_region(
-                HeapObject::LString {
-                    s: slice,
-                    traits: Value::NIL,
-                },
-                region,
-            )
-        }
-        HeapObject::LBytes { data, .. } => {
-            let slice = heap.alloc_region_slice_in_region(data.as_slice(), region);
-            heap.alloc_in_region(
-                HeapObject::LBytes {
-                    data: slice,
-                    traits: Value::NIL,
-                },
-                region,
-            )
-        }
-        HeapObject::LArray { elements, .. } => {
-            let mut copies = Vec::with_capacity(elements.len());
-            for &el in elements.iter() {
-                copies.push(copy_value(heap, region, el, walk)?);
-            }
-            let slice = heap.alloc_region_slice_in_region(&copies, region);
-            heap.alloc_in_region(
-                HeapObject::LArray {
-                    elements: slice,
-                    traits: Value::NIL,
-                },
-                region,
-            )
-        }
-        HeapObject::Float(f) => heap.alloc_in_region(HeapObject::Float(*f), region),
-        other => {
-            return Err(ImageError::Unsupported(format!(
-                "{:?} is not sealed data (docs/impl/image.md § Sealing)",
-                other.tag()
-            )))
-        }
-    };
-    walk.visited.insert(key, copy);
-    Ok(copy)
-}
